# packages/sdk-ts/openapi.json
{"components":{"schemas":{"AGUIChatRequest":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/ChatMsg"},"type":"array"},"run_id":{"type":["string","null"]},"thread_id":{"type":["string","null"]},"tools":{"items":{"$ref":"#/components/schemas/ToolDef"},"type":"array"}},"required":["messages","tools"],"type":"object"},"AgentPart":{"properties":{"id":{"type":"string"},"messageId":{"type":"string"},"name":{"type":"string"},"sessionId":{"type":"string"},"source":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/AgentPartSource"}]},"type":{"type":"string"}},"required":["id","sessionId","messageId","type","name"],"type":"object"},"AgentPartSource":{"properties":{"end":{"minimum":0,"type":"integer"},"start":{"minimum":0,"type":"integer"},"value":{"type":"string"}},"required":["value","start","end"],"type":"object"},"ChatMsg":{"properties":{"content":{},"id":{"type":["string","null"]},"name":{"type":["string","null"]},"role":{"type":"string"},"tool_call_id":{"type":["string","null"]},"tool_calls":{"items":{"$ref":"#/components/schemas/ToolCallWire"},"type":["array","null"]}},"required":["role"],"type":"object"},"FilePart":{"properties":{"filename":{"type":["string","null"]},"id":{"type":"string"},"messageId":{"type":"string"},"mime":{"type":"string"},"sessionId":{"type":"string"},"source":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/FilePartSource"}]},"type":{"type":"string"},"url":{"type":"string"}},"required":["id","sessionId","messageId","type","mime","url"],"type":"object"},"FilePartSource":{"properties":{"path":{"type":"string"},"text":{"$ref":"#/components/schemas/PartTextRange"},"type":{"type":"string"}},"required":["text","type","path"],"type":"object"},"HealthResponse":{"properties":{"status":{"type":"string"},"tool_count":{"minimum":0,"type":"integer"}},"required":["status","tool_count"],"type":"object"},"PartTextRange":{"properties":{"end":{"minimum":0,"type":"integer"},"start":{"minimum":0,"type":"integer"},"value":{"type":"string"}},"required":["value","start","end"],"type":"object"},"PartTime":{"properties":{"end":{"format":"double","type":["number","null"]},"start":{"format":"double","type":"number"}},"required":["start"],"type":"object"},"TextPart":{"properties":{"id":{"type":"string"},"ignored":{"type":["boolean","null"]},"messageId":{"type":"string"},"metadata":{},"sessionId":{"type":"string"},"synthetic":{"type":["boolean","null"]},"text":{"type":"string"},"time":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/PartTime"}]},"type":{"type":"string"}},"required":["id","sessionId","messageId","type","text"],"type":"object"},"ToolCallFuncWire":{"properties":{"arguments":{"type":"string"},"name":{"type":"string"}},"required":["name","arguments"],"type":"object"},"ToolCallRequest":{"properties":{"arguments":{}},"required":["arguments"],"type":"object"},"ToolCallWire":{"properties":{"function":{"$ref":"#/components/schemas/ToolCallFuncWire"},"id":{"type":"string"},"type":{"type":"string"}},"required":["id","type","function"],"type":"object"},"ToolDef":{"properties":{"description":{"type":"string"},"name":{"type":"string"},"parameters":{}},"required":["name","description","parameters"],"type":"object"},"ToolInfo":{"properties":{"description":{"type":"string"},"name":{"type":"string"},"parameters":{}},"required":["name","description","parameters"],"type":"object"},"ToolResult":{"properties":{"content":{"type":"string"},"is_error":{"type":"boolean"},"metadata":{}},"required":["content","is_error"],"type":"object"}}},"info":{"description":"aaBot HTTP/WS API 服务","license":{"identifier":"MIT","name":"MIT"},"title":"aa-server","version":"0.1.0"},"openapi":"3.1.0","paths":{"/health":{"get":{"operationId":"health","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}},"description":"Server health status"}},"tags":["aaBot"]}},"/tools":{"get":{"operationId":"list_tools","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ToolInfo"},"type":"array"}}},"description":"List all available tools"}},"tags":["aaBot"]}},"/tools/{name}":{"post":{"operationId":"call_tool","parameters":[{"description":"Tool name","in":"path","name":"name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolCallRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolResult"}}},"description":"Tool execution result"},"404":{"content":{"text/plain":{"schema":{"type":"string"}}},"description":"Tool not found"},"500":{"content":{"text/plain":{"schema":{"type":"string"}}},"description":"Tool execution failed"}},"tags":["aaBot"]}}},"tags":[{"description":"aaBot API","name":"aaBot"}]}